# swagger.json
{"openapi":"3.0.0","info":{"title":"Tartinchis","version":"1.0.0","description":"This is a simple CRUD API of Tartichis","license":{"name":"MIT","url":"http://mit.com"},"contact":{"name":"Los Chichos","url":"https://github.com/loschichos","email":"chichos@example.com"}},"server":[{"url":"http://localhost:3000"}],"paths":{"/classroom":{"get":{"summary":"Get all classrooms","tags":["Classroom"],"parameters":[{"in":"query","name":"page","schema":{"type":"integer"},"description":"The page number"},{"in":"query","name":"limit","schema":{"type":"integer"},"description":"The number of items to return"}],"responses":{"200":{"description":"The list of classrooms","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Classroom"},"pagination":{"$ref":"#/components/schemas/Pagination"}}}}}}},"400":{"description":"Invalid page or limit parameter","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"post":{"summary":"Create a new classroom","tags":["Classroom"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Classroom"}}}},"responses":{"201":{"description":"The classroom was created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Classroom"}}}}}}},"/classroom/{id}":{"get":{"summary":"Get an classroom by ID","tags":["Classroom"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The classroom ID"}],"responses":{"200":{"description":"The classroom info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Classroom"}}}},"404":{"description":"Classroom not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"delete":{"summary":"Delete an classroom by ID","tags":["Classroom"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The classroom ID"}],"responses":{"200":{"description":"The classroom was successfully deleted","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Classroom"}}}},"404":{"description":"The classroom was not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"put":{"summary":"Update an classroom by ID","tags":["Classroom"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The classroom ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Classroom"}}}},"responses":{"200":{"description":"The classroom was successfully updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Classroom"}}}},"400":{"description":"Some parameters are missing or invalid","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"The classroom was not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/classroom/name/{name}":{"get":{"summary":"Get an classroom by name","tags":["Classroom"],"parameters":[{"in":"path","name":"name","schema":{"type":"string"},"required":true,"description":"The classroom name"}],"responses":{"200":{"description":"The classroom info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Classroom"}}}},"404":{"description":"Classroom not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/classroom/login":{"post":{"summary":"Login as an classroom","tags":["Classroom"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}}},"responses":{"200":{"description":"Successful login","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Classroom"}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/classroom/image-upload":{"post":{"summary":"Upload a image for a classroom","tags":["Classroom"],"consumes":["multipart/form-data"],"parameters":[{"in":"formData","name":"image","type":"file","description":"The file to upload."},{"in":"formData","name":"classroomId","type":"string","description":"The id of the classroom"}],"responses":{"200":{"description":"The image was uploaded successfully"},"404":{"description":"The classroom was not found"}}}},"/team":{"get":{"summary":"Get all teams","tags":["Teams"],"parameters":[{"in":"query","name":"page","schema":{"type":"integer"},"description":"The page number"},{"in":"query","name":"limit","schema":{"type":"integer"},"description":"The number of items to return"}],"responses":{"200":{"description":"The list of teams","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Teams"},"pagination":{"$ref":"#/components/schemas/Pagination"}}}}}}},"400":{"description":"Invalid page or limit parameter","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"post":{"summary":"Create a new team","tags":["Teams"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Teams"}}}},"responses":{"201":{"description":"The team was created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Teams"}}}}}}},"/team/{id}":{"get":{"summary":"Get an team by ID","tags":["Teams"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The team ID"}],"responses":{"200":{"description":"The team info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Teams"}}}},"404":{"description":"Teams not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"delete":{"summary":"Delete an team by ID","tags":["Teams"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The team ID"}],"responses":{"200":{"description":"The team was successfully deleted","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Teams"}}}},"404":{"description":"The team was not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"put":{"summary":"Update an team by ID","tags":["Teams"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The team ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Teams"}}}},"responses":{"200":{"description":"The team was successfully updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Teams"}}}},"400":{"description":"Some parameters are missing or invalid","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"The team was not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/team/name/{name}":{"get":{"summary":"Get an team by name","tags":["Teams"],"parameters":[{"in":"path","name":"name","schema":{"type":"string"},"required":true,"description":"The team name"}],"responses":{"200":{"description":"The team info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Teams"}}}},"404":{"description":"Teams not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/team/login":{"post":{"summary":"Login as an team","tags":["Teams"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}}},"responses":{"200":{"description":"Successful login","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Teams"}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/team/image-upload":{"post":{"summary":"Upload a image for a team","tags":["Teams"],"consumes":["multipart/form-data"],"parameters":[{"in":"formData","name":"image","type":"file","description":"The file to upload."},{"in":"formData","name":"teamId","type":"string","description":"The id of the team"}],"responses":{"200":{"description":"The image was uploaded successfully"},"404":{"description":"The team was not found"}}}},"/user":{"get":{"summary":"Get all users","tags":["User"],"parameters":[{"in":"query","name":"page","schema":{"type":"integer"},"description":"The page number"},{"in":"query","name":"limit","schema":{"type":"integer"},"description":"The number of items to return"}],"responses":{"200":{"description":"The list of users","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/User"},"pagination":{"$ref":"#/components/schemas/Pagination"}}}}}}},"400":{"description":"Invalid page or limit parameter","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"post":{"summary":"Create a new user","tags":["User"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"201":{"description":"The user was created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/user/{id}":{"get":{"summary":"Get an user by ID","tags":["User"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The user ID"}],"responses":{"200":{"description":"The user info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"delete":{"summary":"Delete an user by ID","tags":["User"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The user ID"}],"responses":{"200":{"description":"The user was successfully deleted","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"404":{"description":"The user was not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"put":{"summary":"Update an user by ID","tags":["User"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The user ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"200":{"description":"The user was successfully updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"description":"Some parameters are missing or invalid","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"The user was not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/user/name/{name}":{"get":{"summary":"Get an user by name","tags":["User"],"parameters":[{"in":"path","name":"name","schema":{"type":"string"},"required":true,"description":"The user name"}],"responses":{"200":{"description":"The user info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/user/login":{"post":{"summary":"Login as an user","tags":["User"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}}},"responses":{"200":{"description":"Successful login","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/user/image-upload":{"post":{"summary":"Upload a image for a user","tags":["User"],"consumes":["multipart/form-data"],"parameters":[{"in":"formData","name":"image","type":"file","description":"The file to upload."},{"in":"formData","name":"userId","type":"string","description":"The id of the user"}],"responses":{"200":{"description":"The image was uploaded successfully"},"404":{"description":"The user was not found"}}}}},"components":{"schemas":{"Error":{"type":"object","properties":{"error":{"type":"string","description":"The error message"}}},"MatchCreate":{"type":"object","required":["date","localTeam","visitorTeam","played","round"],"properties":{"date":{"type":"string","format":"date-time","description":"Fecha del partido."},"localTeam":{"$ref":"#/components/schemas/Team","description":"Equipo local."},"visitorTeam":{"$ref":"#/components/schemas/Team","description":"Equipo visitante."},"goalsLocal":{"type":"array","items":{"$ref":"#/components/schemas/User"},"description":"Goles del equipo local."},"goalsVisitor":{"type":"array","items":{"$ref":"#/components/schemas/User"},"description":"Goles del equipo visitante."},"played":{"type":"boolean","description":"Indica si el partido ha sido jugado."},"round":{"type":"number","description":"Número de la ronda.","minimum":1}},"example":{"date":"2023-06-18T18:00:00Z","localTeam":{"_id":"60c84e71ebeb8f001ff60999","name":"Equipo A","initials":"EQA","image":"https://example.com/teamA.png"},"visitorTeam":{"_id":"60c84e71ebeb8f001ff60998","name":"Equipo B","initials":"EQB","image":"https://example.com/teamB.png"},"goalsLocal":[],"goalsVisitor":[],"played":false,"round":1}},"Match":{"allOf":[{"$ref":"#/components/schemas/MatchCreate"},{"type":"object","properties":{"_id":{"type":"string","description":"ID del partido."}},"example":{"_id":"60c84e71ebeb8f001ff60997","date":"2023-06-18T18:00:00Z","localTeam":{"_id":"60c84e71ebeb8f001ff60999","name":"Equipo A","initials":"EQA","image":"https://example.com/teamA.png"},"visitorTeam":{"_id":"60c84e71ebeb8f001ff60998","name":"Equipo B","initials":"EQB","image":"https://example.com/teamB.png"},"goalsLocal":[],"goalsVisitor":[],"played":false,"round":1}}]},"Pagination":{"type":"object","properties":{"totalItems":{"type":"number","description":"Total items on this endpoint"},"totalPages":{"type":"number","description":"Number of pages for this pagination"},"currentPage":{"type":"number","description":"Current page showed"}}},"TeamCreate":{"type":"object","properties":{"name":{"type":"string","description":"Nombre del equipo.","minLength":5,"maxLength":30},"initials":{"type":"string","description":"Iniciales del equipo.","minLength":3,"maxLength":3},"image":{"type":"string","description":"URL de la imagen del equipo."}},"example":{"name":"Equipo A","initials":"EQA","image":"https://example.com/teamA.png"}},"Team":{"allOf":[{"$ref":"#/components/schemas/TeamCreate"},{"type":"object","properties":{"_id":{"type":"string","description":"ID del equipo."}},"example":{"_id":"60c84e71ebeb8f001ff60999","name":"Equipo A","initials":"EQA","image":"https://example.com/teamA.png"}}]},"User":{"type":"object","required":["firstName","lastName","email","password","rol"],"properties":{"_id":{"type":"string","description":"ID del usuario."},"firstName":{"type":"string","description":"Nombre del usuario.","minLength":3,"maxLength":22},"lastName":{"type":"string","description":"Apellido del usuario.","minLength":3,"maxLength":22},"email":{"type":"string","description":"Email del usuario.","format":"email"},"password":{"type":"string","description":"Contraseña del usuario.","minLength":8},"rol":{"type":"string","description":"Rol del usuario.","enum":["PLAYER","MANAGER","ADMIN"]},"team":{"type":"string","description":"ID del equipo al que pertenece el usuario."},"image":{"type":"string","description":"URL de la imagen del usuario."}},"example":{"_id":"60c84e71ebeb8f001ff60999","firstName":"John","lastName":"Doe","email":"john.doe@example.com","password":["hidden"],"rol":"PLAYER","team":"60c84e71ebeb8f001ff60998","image":"https://example.com/johndoe.png"}}}},"tags":[{"name":"Classroom","description":"API for managing classrooms"},{"name":"Teams","description":"API for managing teams"},{"name":"User","description":"API for managing users"}]}